+++ apps/front/src/components/Button/LeaveButton.tsx
